[add] Kegg Service API
[refactoring] Put old Kegg code as deprecated
