+++ src/main/java/com/silicolife/textmining/processes/resources/dictionary/loaders/kegg/webservice/KeggEntitiesEnum.java
@@ -1,0 +1,5 @@
+package com.silicolife.textmining.processes.resources.dictionary.loaders.kegg.webservice;
+
+public enum KeggEntitiesEnum {
+	cpd
+}
